+++ src/helpers/useAuthQuery.ts
@@ -1,0 +1,14 @@
+import { useQuery } from '@apollo/client';
+import { usePrivy } from '@privy-io/react-auth';
+
+export function useAuthQuery(query, options = {}) {
+  const { user } = usePrivy();
+
+  return useQuery(query, {
+    ...options,
+    variables: {
+      privyId: user?.id,
+    },
+    skip: !user,
+  });
+}
